Split clocking out of RCC
